fix(get_transaction, get_data, is_confirmed): put into right places
move verify back, add get_data and is_confirmed functions in utils.py
